+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - US9: SearchPage
 - BUG3: Error during login
 - BUG4: Overflow in login Screen
+- BUG9: Ads photo url are not upload to algolia.